feat: add DocsSearchBar
