+++ src/domain/props/DocsSearchBarProps.ts
@@ -1,0 +1,3 @@
+export interface DocsSearchBarProps {
+  setSearchWords: (searchWords: string[]) => void;
+}
